requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
httplib2>=0.22.0
pywin32>=306
aiofiles==23.2.1
aiohttp==3.9.1
asyncio==3.4.3
beautifulsoup4==4.12.2
cryptography==41.0.7
gspread==5.12.4
httpx==0.25.2
requests==2.31.0
vdf==3.4
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0